* axUtils: cleanup + core/axDetectWine.h: small header with the function of the same name that attempts to detect wine. (could be useful as the wine project does has bugs and breaks things) * axDebugConsole: attempts to check with axDetectWine to put the console window as topmost.

+++ ax/core/axMalloc.h
@@ -147,7 +147,7 @@ TODO:
   {
     static long g_pagesize;
     static long g_regionsize;
-    int rc = NULL;
+    long rc = NULL;
     //slwait(&g_sl);
     while ( InterlockedCompareExchange( &g_sl, 1, 0) != 0 )
       Sleep(0);
